Split `Ctx` into `Ctx` and `CtxHandle` (#112)
The `Ctx` is slit into two parts, the actual `Ctx` and `CtxHandle`
which are meant to be used by sockets. The principal distinction is
that the `Ctx` implements Drop (which terminates the context) while
`CtxHandle` does not. This is convenient because this allows the
lifetime of a `Ctx` to be more obvious (the context is dropped
when `Ctx` is dropped. This also allows `CtxHandle` which implements
copy, which makes usage more ergonomic.

* Renamed `CtxTerminated` to `CtxInvalid`